# openapi.yaml
openapi: 3.1.0
info:
  title: BUILDYNOTE MCP REST API
  description: BUILDYNOTEの仕事・工程・個人予定をAIエージェントから操作するRESTラッパー
  version: 1.0.0
servers:
  - url: https://buildynote-mcp-production.up.railway.app
    description: 本番環境
security:
  - ApiToken: []
components:
  securitySchemes:
    ApiToken:
      type: apiKey
      in: header
      name: X-Api-Token
      description: BUILDYNOTEのAPIトークン（各ユーザーが自分のトークンを使用）
  schemas:
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
        timestamp:
          type: string
          format: date-time
...
paths:
...
  /works:
    get:
      operationId: listWorks
      summary: 仕事一覧取得（ids指定時は詳細を一括取得）
      parameters:
        - name: ids
          in: query
          schema: { type: string }
          description: 仕事IDのカンマ区切り。指定すると複数の仕事詳細を一括取得（works_info相当）
        - name: name
          in: query
          schema: { type: string }
          description: 仕事名（部分一致）
        - name: status
          in: query
          schema: { type: string }
          description: ステータス（1=見込客, 2=受注, 3=完了）
        - name: customer_id
          in: query
          schema: { type: string }
        - name: construction_type_id
          in: query
          schema: { type: string }
          description: 物件区分ID（例:311=Buildynoteシステム開発#issue）。master_construction_typesで一覧取得可能
        - name: sort
          in: query
          schema: { type: string, enum: [desc, updated_desc] }
          description: descまたはupdated_descを指定すると更新日の新しい順で返す
        - name: limit
          in: query
          schema: { type: string }
          description: 取得件数（デフォルト50、最大1000）
        - name: page
          in: query
          schema: { type: string }
          description: ページ番号
      responses:
        '200':
          description: 仕事一覧
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessResponse' }
        '401':
          description: 認証エラー
    post:
      operationId: createWork
      summary: 仕事作成
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, kana, customer_id, construction_type_id, sales_staff_id]
              properties:
                name: { type: string, description: '仕事名（必須）' }
                kana: { type: string, description: '仕事名ふりがな（必須）' }
                customer_id: { type: string }
                construction_type_id: { type: string }
                sales_staff_id: { type: string }
                manager_id: { type: string }
                overview: { type: string, description: '概要（GitLab URL等）' }
                foreign_id: { type: string, description: '外部連携ID' }
                status: { type: string, default: '1' }
      responses:
        '200':
          description: 作成された仕事
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessResponse' }
  /works/{id}:
    get:
      operationId: getWork
      summary: 仕事詳細取得
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: 仕事詳細
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessResponse' }
    put:
      operationId: editWork
      summary: 仕事編集
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, kana]
              properties:
                name: { type: string, description: '仕事名（必須）' }
                kana: { type: string, description: '仕事名ふりがな（必須）' }
                status: { type: string }
                overview: { type: string }
      responses:
        '200':
          description: 編集後の仕事
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessResponse' }
    delete:
      operationId: deleteWork
      summary: 仕事削除
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: 削除結果
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessResponse' }
  /works/{work_id}/gantt:
    get:
      operationId: listGantts
      summary: 工程一覧取得（日付未指定は今日のみ）
      parameters:
        - name: work_id
          in: path
          required: true
          schema: { type: string }
        - name: ids
          in: query
          schema: { type: string }
          description: 工程ID（schedule_id）のカンマ区切り。指定すると複数の工程詳細を一括取得
        - name: start_date
          in: query
          schema: { type: string }
          description: 取得開始日時（YYYY-MM-DDTHH:mm:ss）。省略時は今日のみ
        - name: end_date
          in: query
          schema: { type: string }
          description: 取得終了日時（YYYY-MM-DDTHH:mm:ss）
      responses:
        '200':
          description: 工程一覧
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessResponse' }
    put:
      operationId: editGanttsMulti
      summary: 複数工程の一括編集
      parameters:
        - name: work_id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status, gantts]
              properties:
                status: { type: string, description: '1=公開保存/2=下書き/3=全体調整中/4=停止中' }
                gantts:
                  type: array
                  items:
                    type: object
                    properties:
                      schedule_id: { type: string }
                      name: { type: string }
                      day_start: { type: string }
                      day_end: { type: string }
                      delete_flag: { type: string }
      responses:
        '200':
          description: 編集結果
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessResponse' }
    post:
      operationId: createGantt
      summary: 工程作成
      parameters:
        - name: work_id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, category, day_start, day_end, status]
              properties:
                name: { type: string }
                category: { type: string, description: '1=社内/2=工事/3=納材/4=検査（必須）' }
                day_start: { type: string, format: date }
                day_end: { type: string, format: date }
                status: { type: string, description: '工程状態（1=公開保存, 2=下書き, 3=全体調整中, 4=停止中）。新規作成時は通常 1。必須' }
                industry_type_id: { type: string }
                supplier_company_id: { type: string }
                report_type: { type: string }
      responses:
        '200':
          description: 作成された工程
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessResponse' }
  /works/{work_id}/gantt/{id}:
    get:
      operationId: getGantt
      summary: 工程詳細取得
      parameters:
        - name: work_id
          in: path
          required: true
          schema: { type: string }
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: 工程詳細
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessResponse' }
    put:
      operationId: editGantt
      summary: 工程編集
      parameters:
        - name: work_id
          in: path
          required: true
          schema: { type: string }
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                status: { type: string, description: '工程状態（1=公開保存, 2=下書き, 3=全体調整中, 4=停止中）。省略時は1' }
                name: { type: string }
                day_start: { type: string, format: date }
                day_end: { type: string, format: date }
      responses:
        '200':
          description: 編集後の工程
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessResponse' }
  /schedules:
    get:
      operationId: listSchedules
      summary: 個人予定一覧取得（ids指定時は詳細を一括取得）
      parameters:
        - name: ids
          in: query
          schema: { type: string }
          description: 予定IDのカンマ区切り。指定すると複数の予定詳細を一括取得（user_idは不要になる）
        - name: user_id
          in: query
          schema: { type: string }
          description: 取得対象のユーザーID（listStaffで取得可能）。ids未指定時に必須
        - name: start_date
          in: query
          schema: { type: string }
          description: 取得開始日時（YYYY-MM-DD または YYYY-MM-DDTHH:mm:ss）
        - name: end_date
          in: query
          schema: { type: string }
          description: 取得終了日時（YYYY-MM-DD または YYYY-MM-DDTHH:mm:ss）。start_dateと合わせて範囲指定
        - name: work_id
          in: query
          schema: { type: string }
      responses:
        '200':
          description: 予定一覧
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessResponse' }
    post:
      operationId: createSchedule
      summary: 個人予定作成
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, start_date, end_date, user_list]
              properties:
                name: { type: string }
                start_date: { type: string, description: '開始日（YYYY-MM-DD）' }
                start_time: { type: string, description: '開始時刻（HH:mm）。省略可' }
                end_date: { type: string, description: '終了日（YYYY-MM-DD）' }
                end_time: { type: string, description: '終了時刻（HH:mm）。省略可' }
                user_list: { type: array, items: { type: object, properties: { user_id: { type: string } } }, description: '参加者リスト（例: [{user_id: "277"}]）。削除したい場合はAPIユーザー({user_id: "8497"})も含める' }
                work_id: { type: string }
                is_private: { type: string, description: '0=公開, 1=非公開' }
                label_id: { type: string }
      responses:
        '200':
          description: 作成された予定
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessResponse' }
  /schedules/{id}:
    get:
      operationId: getSchedule
      summary: 個人予定詳細取得
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: 予定詳細
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessResponse' }
    put:
      operationId: editSchedule
      summary: 個人予定編集
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                start_date: { type: string, description: '開始日（YYYY-MM-DD）' }
                start_time: { type: string, description: '開始時刻（HH:mm）。省略可' }
                end_date: { type: string, description: '終了日（YYYY-MM-DD）' }
                end_time: { type: string, description: '終了時刻（HH:mm）。省略可' }
                user_list: { type: array, items: { type: object, properties: { user_id: { type: string } } } }
                work_id: { type: string }
      responses:
        '200':
          description: 編集後の予定
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessResponse' }
    delete:
      operationId: deleteSchedule
      summary: 個人予定削除
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: 削除結果
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessResponse' }
  /masters/staff:
    get:
      operationId: listStaff
      summary: 社員一覧（ids指定時は詳細を一括取得）
      description: |
        idsクエリを省略した場合は全社員のID・名前のみを返す（軽量）。
        idsを指定した場合は該当社員の詳細（支社・部署・役職を含む）を一括取得する。
      parameters:
        - name: ids
          in: query
          required: false
          description: 社員IDのカンマ区切り（例 1,3,5）。指定すると詳細情報を一括取得
          schema: { type: string }
      responses:
        '200':
          description: 社員一覧または詳細
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessResponse' }
  /masters/staff/{id}:
    get:
      operationId: getStaffInfo
      summary: 社員詳細（支社・部署・役職含む）
      description: 指定した社員IDの詳細情報を取得する。listStaffに含まれない支社・部署・役職・メール・電話などを返す。
      parameters:
        - name: id
          in: path
          required: true
          description: 社員ID
          schema: { type: string }
      responses:
        '200':
          description: 社員詳細
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessResponse' }
  /masters/construction-types:
    get:
      operationId: listConstructionTypes
      summary: 物件区分一覧（ids指定時は詳細を一括取得）
      parameters:
        - name: ids
          in: query
          required: false
          schema: { type: string }
          description: 物件区分IDのカンマ区切り
      responses:
        '200':
          description: 物件区分一覧
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessResponse' }
    post:
      operationId: createConstructionType
      summary: 物件区分を新規作成
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                foreign_id: { type: string }
      responses:
        '200':
          description: 作成結果
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessResponse' }
  /masters/construction-types/{id}:
    get:
      operationId: getConstructionType
      summary: 物件区分の詳細を取得
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: 物件区分詳細
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessResponse' }
    put:
      operationId: editConstructionType
      summary: 物件区分を編集
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                foreign_id: { type: string }
      responses:
        '200':
          description: 編集結果
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessResponse' }
    delete:
      operationId: deleteConstructionType
      summary: 物件区分を削除
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: 削除結果
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessResponse' }
  /masters/industry-types:
    get:
      operationId: listIndustryTypes
      summary: 業種区分一覧（ids指定時は詳細を一括取得）
      parameters:
        - name: ids
          in: query
          required: false
          schema: { type: string }
      responses:
        '200':
          description: 業種区分一覧
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessResponse' }
    post:
      operationId: createIndustryType
      summary: 業種区分を新規作成
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                foreign_id: { type: string }
      responses:
        '200':
          description: 作成結果
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessResponse' }
  /masters/industry-types/{id}:
    get:
      operationId: getIndustryType
      summary: 業種区分の詳細を取得
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: 業種区分詳細
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessResponse' }
    put:
      operationId: editIndustryType
      summary: 業種区分を編集
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                foreign_id: { type: string }
      responses:
        '200':
          description: 編集結果
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessResponse' }
    delete:
      operationId: deleteIndustryType
      summary: 業種区分を削除
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: 削除結果
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessResponse' }
  /masters/customers:
    get:
      operationId: listCustomers
      summary: 顧客一覧（ids指定時は詳細を一括取得）
      parameters:
        - name: ids
          in: query
          required: false
          schema: { type: string }
        - name: page
          in: query
          required: false
          schema: { type: string }
        - name: limit
          in: query
          required: false
          schema: { type: string }
      responses:
        '200':
          description: 顧客一覧
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessResponse' }
    post:
      operationId: createCustomer
      summary: 顧客を新規作成
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, kana]
              properties:
                name: { type: string }
                kana: { type: string }
                postcode: { type: string }
                prefecture: { type: string }
                city: { type: string }
                town: { type: string }
                tel: { type: string }
                mobile: { type: string }
                email: { type: string }
                foreign_id: { type: string }
                memo: { type: string }
      responses:
        '200':
          description: 作成結果
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessResponse' }
...
  /masters/offices:
    get:
      operationId: listOffices
      summary: 支社一覧（ids指定時は詳細を一括取得）
      parameters:
        - name: ids
          in: query
          required: false
          schema: { type: string }
      responses:
        '200':
          description: 支社一覧
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessResponse' }
    post:
      operationId: createOffice
      summary: 支社を新規作成
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, kana, postcode, prefecture, city, tel]
              properties:
                name: { type: string }
                kana: { type: string }
                postcode: { type: string }
                prefecture: { type: string }
                city: { type: string }
                town: { type: string }
                tel: { type: string }
                fax: { type: string }
                foreign_id: { type: string }
      responses:
        '200':
          description: 作成結果
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessResponse' }
...
  /masters/departments:
    get:
      operationId: listDepartments
      summary: 部署一覧（ids指定時は詳細を一括取得）
      parameters:
        - name: ids
          in: query
          required: false
          schema: { type: string }
      responses:
        '200':
          description: 部署一覧
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessResponse' }
    post:
      operationId: createDepartment
      summary: 部署を新規作成
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                foreign_id: { type: string }
      responses:
        '200':
          description: 作成結果
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessResponse' }
...
  /masters/positions:
    get:
      operationId: listPositions
      summary: 役職一覧（ids指定時は詳細を一括取得）
      parameters:
        - name: ids
          in: query
          required: false
          schema: { type: string }
      responses:
        '200':
          description: 役職一覧
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessResponse' }
    post:
      operationId: createPosition
      summary: 役職を新規作成
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                foreign_id: { type: string }
      responses:
        '200':
          description: 作成結果
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessResponse' }
...
  /masters/property-types:
    get:
      operationId: listPropertyTypes
      summary: 建物タイプ一覧（ids指定時は詳細を一括取得）
      parameters:
        - name: ids
          in: query
          required: false
          schema: { type: string }
      responses:
        '200':
          description: 建物タイプ一覧
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessResponse' }
    post:
      operationId: createPropertyType
      summary: 建物タイプを新規作成
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                foreign_id: { type: string }
      responses:
        '200':
          description: 作成結果
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessResponse' }
...
  /masters/companies:
    get:
      operationId: listCompanies
      summary: 協力会社一覧（ids指定時は詳細を一括取得、読み取り専用）
      parameters:
        - name: ids
          in: query
          required: false
          schema: { type: string }
      responses:
        '200':
          description: 協力会社一覧
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessResponse' }
...